+++ test/PHPMailer/ReplyToGetSetClearTest.php
@@ -1,0 +1,121 @@
+<?php
+
+/**
+ * PHPMailer - PHP email transport unit tests.
+ * PHP version 5.5.
+ *
+ * @author    Marcus Bointon <phpmailer@synchromedia.co.uk>
+ * @author    Andy Prevost
+ * @copyright 2012 - 2020 Marcus Bointon
+ * @copyright 2004 - 2009 Andy Prevost
+ * @license   http://www.gnu.org/copyleft/lesser.html GNU Lesser General Public License
+ */
+
+namespace PHPMailer\Test\PHPMailer;
+
+use PHPMailer\PHPMailer\PHPMailer;
+use PHPMailer\Test\SendTestCase;
+
+/**
+ * Test reply-to address setting, getting and clearing functionality.
+ */
+final class ReplyToGetSetClearTest extends SendTestCase
+{
+
+    /**
+     * Test low priority.
+     */
+    public function testLowPriority()
+    {
+        $this->Mail->Body = 'Here is the main body.  There should be ' .
+            'a reply to address in this message.';
+        $this->Mail->Subject .= ': Low Priority';
+        $this->Mail->addReplyTo('nobody@nobody.com', 'Nobody (Unit Test)');
+
+        $this->buildBody();
+        self::assertTrue($this->Mail->send(), $this->Mail->ErrorInfo);
+    }
+
+    /**
+     * Test addressing.
+     */
+    public function testAddressing()
+    {
+        self::assertTrue($this->Mail->addReplyTo('a@example.com'), 'Replyto Addressing failed');
+        self::assertFalse($this->Mail->addReplyTo('a@example..com'), 'Invalid Replyto address accepted');
+        $this->Mail->clearReplyTos();
+    }
+
+    /**
+     * Tests CharSet and Unicode -> ASCII conversions for addresses with IDN.
+     */
+    public function testConvertEncoding()
+    {
+        if (!PHPMailer::idnSupported()) {
+            self::markTestSkipped('intl and/or mbstring extensions are not available');
+        }
+
+        $this->Mail->clearReplyTos();
+
+        //This file is UTF-8 encoded. Create a domain encoded in "iso-8859-1".
+        $letter = html_entity_decode('&ccedil;', ENT_COMPAT, PHPMailer::CHARSET_ISO88591);
+        $domain = '@' . 'fran' . $letter . 'ois.ch';
+        $this->Mail->addReplyTo('test+replyto' . $domain);
+
+        //Queued addresses are not returned by get*Addresses() before send() call.
+        self::assertEmpty($this->Mail->getReplyToAddresses(), 'Bad "reply-to" recipients');
+
+        $this->buildBody();
+        self::assertTrue($this->Mail->send(), $this->Mail->ErrorInfo);
+
+        //Addresses with IDN are returned by get*Addresses() after send() call.
+        $domain = $this->Mail->punyencodeAddress($domain);
+        self::assertSame(
+            ['test+replyto' . $domain => ['test+replyto' . $domain, '']],
+            $this->Mail->getReplyToAddresses(),
+            'Bad "reply-to" addresses'
+        );
+    }
+
+    /**
+     * Tests removal of duplicate recipients and reply-tos.
+     */
+    public function testDuplicateIDNRemoved()
+    {
+        if (!PHPMailer::idnSupported()) {
+            self::markTestSkipped('intl and/or mbstring extensions are not available');
+        }
+
+        $this->Mail->clearReplyTos();
+
+        $this->Mail->CharSet = PHPMailer::CHARSET_UTF8;
+
+        self::assertTrue($this->Mail->addReplyTo('test+replyto@françois.ch'));
+        self::assertFalse($this->Mail->addReplyTo('test+replyto@françois.ch'));
+        self::assertTrue($this->Mail->addReplyTo('test+replyto@FRANÇOIS.CH'));
+        self::assertFalse($this->Mail->addReplyTo('test+replyto@FRANÇOIS.CH'));
+        self::assertTrue($this->Mail->addReplyTo('test+replyto@xn--franois-xxa.ch'));
+        self::assertFalse($this->Mail->addReplyTo('test+replyto@xn--franois-xxa.ch'));
+        self::assertFalse($this->Mail->addReplyTo('test+replyto@XN--FRANOIS-XXA.CH'));
+
+        $this->buildBody();
+        self::assertTrue($this->Mail->send(), $this->Mail->ErrorInfo);
+
+        //There should be only one "Reply-To" address.
+        self::assertCount(
+            1,
+            $this->Mail->getReplyToAddresses(),
+            'Bad count of "reply-to" addresses'
+        );
+    }
+
+    public function testGivenIdnAddress_addReplyTo_returns_true()
+    {
+        if (file_exists(\PHPMAILER_INCLUDE_DIR . '/test/fakefunctions.php') === false) {
+            $this->markTestSkipped('/test/fakefunctions.php file not found');
+        }
+
+        include \PHPMAILER_INCLUDE_DIR . '/test/fakefunctions.php';
+        $this->assertTrue($this->Mail->addReplyTo('test@françois.ch'));
+    }
+}
